add GameState
GameState references current objects(ptrs). Renderer as friend of gamestate can be used to render the state of the game.
GameState is an observer, but because it contains ptrs, it does not need an onNotify if the object moves etc. Making it an observer simplifies replacement/deletion of objects instead.